LayoutTests/imported/w3c/web-platform-tests/editing/other/paste-multiline-text-as-plaintext.tentative.https_white-space=pre-line-expected.txt

abc
def
ghi



FAIL Inserting multiline text shouldn't be split to multiple Text nodes unless using br elements assert_equals: expected "<p>abc\ndef\nghi</p>" but got "<p><br></p>"
